+++ test/tensorrt/test_converter_activation.py
@@ -72,5 +72,19 @@ class TestTANHTRTPattern(TensorRTBaseTest):
         self.check_trt_result()
 
 
+class TestSigmoidTRTPattern(TensorRTBaseTest):
+    def setUp(self):
+        self.python_api = paddle.nn.functional.sigmoid
+        self.api_args = {
+            "x": np.random.randn(2, 3).astype(np.float32),
+        }
+        self.program_config = {"feed_list": ["x"]}
+        self.min_shape = {"x": [1, 3], "y": [1, 3]}
+        self.max_shape = {"x": [5, 3], "y": [5, 3]}
+
+    def test_trt_result(self):
+        self.check_trt_result()
+
+
 if __name__ == '__main__':
     unittest.main()
